TOUCHFILE to include time stamp - avoid race conditions
It's rather easily possible, that the $TOUCHFILE already exists, maybe from a previous run, or so. Add a timestamp to the filename and a random number and the current pid.

This should make it close to "impossible" that the file exists from previous runs or so.

Would've liked to use mktemp, but I'm unsure how portable "mktemp -u" (do not create anything; merely print a name (unsafe)) would've been.